@@ -1,9 +1,13 @@
 {
-  "name": "@dontcode/seed",
-  "version": "0.0.10",
+  "name": "@dontcode/plugin-seed",
+  "version": "0.1.0",
   "peerDependencies": {
-    "@angular/common": "^12.2.13",
-    "@angular/core": "^12.2.13"
+    "@angular/common": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@dontcode/core": "^0.5.2",
+    "@dontcode/plugin-common": "^0.5.5",
+    "@dontcode/sandbox": "^0.5.5",
+    "rxjs": "^7.5.5"
   },
   "dependencies": {
     "tslib": "^2.3.1"
